% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_HVG_2_Evaluation_Criteria.R
\name{ARI_NMI_F1_func_Max_continuous}
\alias{ARI_NMI_F1_func_Max_continuous}
\title{Criterion 6: NMI}
\usage{
ARI_NMI_F1_func_Max_continuous(embedding, pro)
}
\arguments{
\item{embedding}{embedding}

\item{pro}{pro}

\item{resolution}{resolution}
}
\description{
Criterion 6: NMI
}
\details{
6
}
