% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_HVG_2_Evaluation_Criteria.R
\name{ARI_NMI_func}
\alias{ARI_NMI_func}
\title{Criterion 1: ARI & NMI with closest cell type number}
\usage{
ARI_NMI_func(embedding, cell_label, maxit = 10)
}
\arguments{
\item{embedding}{embedding}

\item{cell_label}{cell_label}

\item{maxit}{max iteration}
}
\description{
Criterion 1: ARI & NMI with closest cell type number
}
\details{
1
}
