% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_HVG_2_Evaluation_Criteria.R
\name{ARILouvain}
\alias{ARILouvain}
\title{Criterion 1: ARI}
\usage{
ARILouvain(embedding, cell_label, maxit = 25)
}
\arguments{
\item{embedding}{embedding}

\item{cell_label}{cell_label}

\item{maxit}{maxit}
}
\description{
Criterion 1: ARI
}
