% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_HVG_2_Evaluation_Criteria.R
\name{asw_func_discrete}
\alias{asw_func_discrete}
\title{Criterion 6: ASW}
\usage{
asw_func_discrete(embedding, cell_label)
}
\arguments{
\item{embedding}{embedding}

\item{cell_label}{cell_label}
}
\description{
Criterion 6: ASW
}
\details{
6
}
