% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_HVG_2_Evaluation_Criteria.R
\name{asw_max_func_continuous}
\alias{asw_max_func_continuous}
\title{Criterion 5: ASW}
\usage{
asw_max_func_continuous(embedding, dmat)
}
\arguments{
\item{embedding}{embedding}

\item{dmat}{dmat}
}
\description{
Criterion 5: ASW
}
\details{
5
}
