% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_HVG_2_Evaluation_Criteria.R
\name{between_within_dist_ratio}
\alias{between_within_dist_ratio}
\title{Criterion 4: Within VS Between Cell Type Distance Ratio}
\usage{
between_within_dist_ratio(embedding, cell_label)
}
\arguments{
\item{embedding}{embedding}

\item{cell_label}{cell_label}
}
\description{
Criterion 4: Within VS Between Cell Type Distance Ratio
}
\details{
4
}
