% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_HVG_2_Evaluation_Criteria.R
\name{evaluate_hvg_continuous}
\alias{evaluate_hvg_continuous}
\title{Comprehensive Evaluation for CITEseq & MultiomeATAC
Select input from "MultiomeATAC" or "CITEseq}
\usage{
evaluate_hvg_continuous(pcalist, pro, input = "MultiomeATAC")
}
\arguments{
\item{pcalist}{pcalist}

\item{pro}{protein}

\item{input}{input type}
}
\value{
var_ratio

knn_ratio

3nn

dist_cor
}
\description{
Comprehensive Evaluation for CITEseq & MultiomeATAC
Select input from "MultiomeATAC" or "CITEseq
}
