% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_HVG_2_Evaluation_Criteria.R
\name{evaluate_hvg_discrete}
\alias{evaluate_hvg_discrete}
\title{Comprehensive Evaluation for cell sorting.}
\usage{
evaluate_hvg_discrete(pcalist, label, verbose = T)
}
\arguments{
\item{pcalist}{pcalist}

\item{label}{label}

\item{verbose}{verbose}
}
\value{
var_ratio

ari

3nn

dist_ratio
}
\description{
Comprehensive Evaluation for cell sorting.
}
\details{
evaluate
}
