% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_HVG_1_Methods.R
\name{hvg_pca}
\alias{hvg_pca}
\title{hvg_pca}
\usage{
hvg_pca(rna_mat, nfeatures = 2000, SCT = TRUE)
}
\arguments{
\item{rna_mat}{input scRNA-seq count matrix}

\item{nfeatures}{Number of features to select as top variable features.}

\item{SCT}{input SCT}
}
\value{
seurat.obj.pca:  a list of PCA.

var.seurat.obj:
}
\description{
hvg_pca
}
\details{
hvg pca
}
\examples{
if(0){
simple_matrix<-matrix(1:2e4,nrow=4000,ncol=5)
rownames(simple_matrix)<-1:nrow(simple_matrix)
colnames(simple_matrix)<-1:ncol(simple_matrix)
simple_matrix_HVG<-FindVariableFeaturesMix(simple_matrix)
}

}
