% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_HVG_2_Evaluation_Criteria.R
\name{knn_accuracy}
\alias{knn_accuracy}
\title{Criterion 3: Nearest Neighbor Accuracy}
\usage{
knn_accuracy(embedding, cell_label, k = 3, cutoff = 3000)
}
\arguments{
\item{embedding}{embedding}

\item{cell_label}{cell_label}

\item{k}{k}

\item{cutoff}{cutoff}
}
\description{
Criterion 3: Nearest Neighbor Accuracy
}
