% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_HVG_2_Evaluation_Criteria.R
\name{knn_ratio}
\alias{knn_ratio}
\title{Criterion 3: Nearest Neighbor Distance Ratio}
\usage{
knn_ratio(embedding, pro, k = 100, cutoff = 5000)
}
\arguments{
\item{embedding}{embedding}

\item{pro}{protein}

\item{k}{k}

\item{cutoff}{cutoff}
}
\description{
Criterion 3: Nearest Neighbor Distance Ratio
}
\details{
3
}
