% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_HVG_2_Evaluation_Criteria.R
\name{knn_regression}
\alias{knn_regression}
\title{Criterion 2: Nearest Neighbor Mean Square Error}
\usage{
knn_regression(embedding, pro, k = 3, cutoff = 5000)
}
\arguments{
\item{embedding}{embedding}

\item{pro}{pro}

\item{k}{k}

\item{cutoff}{cutoff}
}
\description{
Criterion 2: Nearest Neighbor Mean Square Error
}
\details{
2
}
