% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_HVG_2_Evaluation_Criteria.R
\name{lisi_score_func}
\alias{lisi_score_func}
\title{Criterion 5: LISI Score}
\usage{
lisi_score_func(embedding, cell_label)
}
\arguments{
\item{embedding}{embedding}

\item{cell_label}{cell_label}
}
\description{
Criterion 5: LISI Score
}
\details{
5
}
