% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_HVG_1_Methods.R
\name{mixture_hvg_pca0}
\alias{mixture_hvg_pca0}
\title{mixture_hvg_pca0}
\usage{
mixture_hvg_pca0(rna_mat, nfeatures = 2000)
}
\arguments{
\item{rna_mat}{input scRNA-seq count matrix}

\item{nfeatures}{Number of features to select as top variable features.}
}
\value{
seurat.obj.pca:  a list of PCA.

var.seurat.obj:
}
\description{
mixture_hvg_pca0
}
\details{
mixture
}
