% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_HVG_2_Evaluation_Criteria.R
\name{nmi_func}
\alias{nmi_func}
\title{Criterion 6: NMI}
\usage{
nmi_func(embedding, pro, resolution = 0.2)
}
\arguments{
\item{embedding}{embedding}

\item{pro}{pro}
}
\description{
Criterion 6: NMI
}
\details{
6
}
