% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_HVG_2_Evaluation_Criteria.R
\name{within_between_dist_ratio}
\alias{within_between_dist_ratio}
\title{Criterion 4: Within VS Between Cell Type Distance Ratio}
\usage{
within_between_dist_ratio(embedding, cell_label)
}
\arguments{
\item{cell_label}{}
}
\description{
Criterion 4: Within VS Between Cell Type Distance Ratio
}
