% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_HVG_2_Evaluation_Criteria.R
\name{within_between_var_ratio}
\alias{within_between_var_ratio}
\title{Criterion 2: Within VS Between Cell Type Variance Ratio}
\usage{
within_between_var_ratio(embedding, cell_label)
}
\arguments{
\item{embedding}{embedding}

\item{cell_label}{cell_label}
}
\description{
Criterion 2: Within VS Between Cell Type Variance Ratio
}
\details{
2
}
