% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_HVG_2_Evaluation_Criteria.R
\name{within_between_var_ratio_continuous}
\alias{within_between_var_ratio_continuous}
\title{Criterion 1: Variance Ratio}
\usage{
within_between_var_ratio_continuous(embedding, pro, resolution = 0.2)
}
\arguments{
\item{resolution}{}
}
\description{
Criterion 1: Variance Ratio
}
