% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_HVG_2_Evaluation_Criteria.R
\name{within_between_var_ratio_max_continuous}
\alias{within_between_var_ratio_max_continuous}
\title{Criterion 1: Variance Ratio}
\usage{
within_between_var_ratio_max_continuous(embedding, pro)
}
\arguments{
\item{embedding}{embedding}

\item{pro}{pro}
}
\description{
Criterion 1: Variance Ratio
}
\details{
1
}
